+++ linux/history.md
@@ -1,5 +1,8 @@
 # Keyman for Linux Version History
 
+## 2019-02-18 11.0.109 beta
+* create appstream appdata for keyman-config (#1543)
+
 ## 2019-02-15 11.0.108 beta
 * handle keyboard package install/uninstall when languages aren't defined (#1585)