+++ linux/history.md
@@ -6,8 +6,8 @@
 ## 2019-02-20 11.0.110 beta
 * update readme about launchpad (#1574)
 
-## 2019-02-18  11.0.109 beta
-* add appstream appdata file for km-config (#1543)
+## 2019-02-18 11.0.109 beta
+* add appstream appdata file for keyman-config (#1543)
 
 ## 2019-02-15 11.0.108 beta
 * handle keyboard package install/uninstall when languages aren't defined (#1585)